moved files into sub folder and removed indexing artefact
