+++ settings/questionsSelection.ts
@@ -10,10 +10,10 @@ import { AsymptoticsPreciseLanguage } from "@shared/question-generators/asymptot
 import { SortTerms } from "@shared/question-generators/asymptotics/sort"
 import { SimplifySum } from "@shared/question-generators/asymptotics/sum"
 import { huffmanCoding } from "@shared/question-generators/huffman-coding/huffmanCoding"
-import { CRT } from "@shared/question-generators/math/crt"
-import { ExtendedEuclideanAlgorithm } from "@shared/question-generators/math/eea"
-import { modFactor } from "@shared/question-generators/math/modFactorization"
-import { ModTricks } from "@shared/question-generators/math/modTricks"
+import { CRT } from "@shared/question-generators/math/modularArithmetic/crt"
+import { ExtendedEuclideanAlgorithm } from "@shared/question-generators/math/modularArithmetic/eea"
+import { modFactor } from "@shared/question-generators/math/modularArithmetic/modFactorization"
+import { ModTricks } from "@shared/question-generators/math/modularArithmetic/modTricks"
 import { NormalForms } from "@shared/question-generators/propositional-logic/normalForms.ts"
 import { Satisfiability } from "@shared/question-generators/propositional-logic/satisfiability.ts"
 import { queueQuestion } from "@shared/question-generators/Queue/QueueGenerator.ts"
