+++ shared/src/question-generators/math/modularArithmetic/modTricks.test.ts
