+++ shared/src/question-generators/math/modularArithmetic/utils.ts
